feat: 리뷰 반영 - 카테고리 메뉴 리팩토링
